requirements.txt
openai
tiktoken
python_dotenv